Smart-Home-Energy-Saver-AI/requirements.txt
a2a-sdk==0.3.11
ag-ui-protocol==0.1.10
agent-framework==1.0.0b251111
agent-framework-a2a==1.0.0b251111
agent-framework-ag-ui==1.0.0b251111
agent-framework-anthropic==1.0.0b251111
agent-framework-azure-ai==1.0.0b251111
agent-framework-chatkit==1.0.0b251111
agent-framework-copilotstudio==1.0.0b251111
agent-framework-core==1.0.0b251111
agent-framework-devui==1.0.0b251111
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b251111
agent-framework-purview==1.0.0b251111
agent-framework-redis==1.0.0b251111
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b4
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmdstanpy==1.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.1
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.8.4
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
holidays==0.84
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.21.0
mem0ai==1.0.0
microsoft-agents-activity==0.5.3
microsoft-agents-copilotstudio-client==0.5.3
microsoft-agents-hosting-core==0.5.3
ml_dtypes==0.5.3
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
narwhals==2.11.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.7.2
openai-agents==0.5.0
openai-chatkit==1.1.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
optuna==4.6.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
ply==3.11
portalocker==3.2.0
posthog==7.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.15.1
redis==6.4.0
redisvl==0.11.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.49.3
streamlit==1.51.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
